feat(ui): merge Usage tab into Plan & usage settings
Move the spend analysis content from the standalone Usage nav tab into
the Settings → Plan & usage section, remove the sidebar item, and keep
/usage as a redirect so restored windows and stale history still land
on the merged page.

Generated-By: PostHog Code
Task-Id: eed52a43-8346-4e29-a370-e99c6a1a862c
