+++ packages/ui/src/features/command/CommandMenu.tsx
@@ -58,7 +58,6 @@ import {
   navigateToChannel,
   navigateToCommandCenter,
   navigateToInbox,
-  navigateToUsage,
 } from "@posthog/ui/router/navigationBridge";
 import { useAppView } from "@posthog/ui/router/useAppView";
 import { openTask, openTaskInput } from "@posthog/ui/router/useOpenTask";
@@ -280,15 +279,12 @@ export function CommandMenu({ open, onOpenChange }: CommandMenuProps) {
         },
       },
       {
-        id: "usage",
-        label: "Usage",
-        keywords: "billing spend cost credits",
+        id: "plan-usage",
+        label: "Plan & usage",
+        keywords: "billing spend cost credits usage plan",
         icon: <ChartLine size={12} className="text-gray-11" />,
         action: "open-usage",
-        onRun: () => {
-          closeSettingsDialog();
-          navigateToUsage();
-        },
+        onRun: () => openSettingsDialog("plan-usage"),
       },
       {
         id: "go-back",
