+++ packages/ui/src/features/sidebar/components/SidebarNavSection.tsx
@@ -7,7 +7,6 @@ import { useFeatureFlag } from "@posthog/ui/features/feature-flags/useFeatureFla
 import { useInboxAllReports } from "@posthog/ui/features/inbox/hooks/useInboxAllReports";
 import { useSidebarStore } from "@posthog/ui/features/sidebar/sidebarStore";
 import { useTasks } from "@posthog/ui/features/tasks/useTasks";
-import { useSpendAnalysisEnabled } from "@posthog/ui/features/usage/useSpendAnalysisEnabled";
 import {
   navigateToActivity,
   navigateToAgents,
@@ -16,7 +15,6 @@ import {
   navigateToInbox,
   navigateToMcpServers,
   navigateToSkills,
-  navigateToUsage,
   navigateToWebsiteCommandCenter,
   navigateToWebsiteHome,
   navigateToWebsiteMcpServers,
@@ -38,7 +36,6 @@ import { McpServersItem } from "./items/McpServersItem";
 import { NewTaskItem } from "./items/NewTaskItem";
 import { SearchItem } from "./items/SearchItem";
 import { SkillsItem } from "./items/SkillsItem";
-import { UsageItem } from "./items/UsageItem";
 
 const SIDEBAR_INBOX_REFETCH_INTERVAL_MS = 60_000;
 
@@ -56,14 +53,13 @@ interface SidebarNavSectionProps {
 // state, badge count, and click handler is wired here — so it can be dropped
 // into either layout. In the Channels space, destinations with a /website
 // mirror (Home, Skills, MCP servers, Command Center) stay in that space;
-// Inbox, Agents, Usage and New task have no mirror yet and jump back to Code.
+// Inbox, Agents and New task have no mirror yet and jump back to Code.
 // Search opens the command menu in place.
 export function SidebarNavSection({
   commandCenterActiveCount: providedActiveCount,
 }: SidebarNavSectionProps = {}) {
   const view = useAppView();
   const homeTabEnabled = useFeatureFlag(HOME_TAB_FLAG);
-  const usageEnabled = useSpendAnalysisEnabled();
   // Channels stay behind project-bluebird: the "Enable channels" nav row (and
   // the Canvas row it reveals) only appear where the canvas backend is wired.
   const bluebirdEnabled = useFeatureFlag(
@@ -76,8 +72,8 @@ export function SidebarNavSection({
 
   // When this section renders inside the Channels space, the destinations that
   // have a /website mirror stay in that space; everything else (and the whole
-  // section in the Code space) uses the canonical routes. Inbox, Agents, Usage
-  // and New task have no mirror yet, so they intentionally jump back to Code.
+  // section in the Code space) uses the canonical routes. Inbox, Agents and
+  // New task have no mirror yet, so they intentionally jump back to Code.
   const inChannels = useRouterState({
     select: (s) => s.location.pathname.startsWith("/website"),
   });
@@ -103,7 +99,6 @@ export function SidebarNavSection({
   const isCommandCenterActive = view.type === "command-center";
   const isSkillsActive = view.type === "skills";
   const isMcpServersActive = view.type === "mcp-servers";
-  const isUsageActive = view.type === "usage";
 
   // Open pull requests in the inbox — the main CTA, and the same count the inbox
   // Pull requests tab shows, so the badge and the tab always agree.
@@ -174,19 +169,13 @@ export function SidebarNavSection({
         <McpServersItem isActive={isMcpServersActive} onClick={goMcpServers} />
       </Box>
 
-      <Box mb={usageEnabled || bluebirdEnabled ? undefined : "2"}>
+      <Box mb={bluebirdEnabled ? undefined : "2"}>
         <CommandCenterItem
           isActive={isCommandCenterActive}
           onClick={goCommandCenter}
           activeCount={commandCenterActiveCount}
         />
       </Box>
-
-      {usageEnabled && (
-        <Box mb="2">
-          <UsageItem isActive={isUsageActive} onClick={navigateToUsage} />
-        </Box>
-      )}
 
       {/* "Channels" is a toggle laid out as a nav row: the # label and Alpha
           badge on the left, a Switch on the right. It flips the channels
