+++ packages/ui/src/router/navigationBridge.ts
@@ -185,10 +185,6 @@ export function navigateToMcpServers(): void {
   void getRouterOrNull()?.navigate({ to: "/mcp-servers" });
 }
 
-export function navigateToUsage(options?: { replace?: boolean }): void {
-  void getRouterOrNull()?.navigate({ to: "/usage", replace: options?.replace });
-}
-
 // Channels-space mirrors. These render the same shared views as their /code (or
 // top-level) counterparts but under /website, so navigating from the channels
 // sidebar keeps the channels chrome instead of switching back to Code. The
