+++ packages/ui/src/router/routes/usage.tsx
@@ -1,6 +1,14 @@
-import { UsageView } from "@posthog/ui/features/usage/UsageView";
-import { createFileRoute } from "@tanstack/react-router";
+import { createFileRoute, redirect } from "@tanstack/react-router";
 
+// The standalone Usage tab merged into Settings → Plan & usage. Keep the
+// route as a redirect so restored windows and stale history entries land
+// on the merged page instead of a not-found screen.
 export const Route = createFileRoute("/usage")({
-  component: UsageView,
+  beforeLoad: () => {
+    throw redirect({
+      to: "/settings/$category",
+      params: { category: "plan-usage" },
+      replace: true,
+    });
+  },
 });
