+++ packages/ui/src/router/useAppView.ts
@@ -19,7 +19,6 @@ export type AppViewType =
   | "command-center"
   | "skills"
   | "mcp-servers"
-  | "usage"
   | "settings";
 
 export interface AppView {
@@ -84,8 +83,6 @@ function deriveFromMatches(matches: Match[]): AppView {
     case "/mcp-servers":
     case "/website/mcp-servers":
       return { type: "mcp-servers" };
-    case "/usage":
-      return { type: "usage" };
     case "/settings/$category":
     case "/settings/":
       return { type: "settings" };
